% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_tests.R
\name{add_features}
\alias{add_features}
\title{Add Features to Time Series}
\usage{
add_features(data)
}
\arguments{
\item{data}{is an \code{xts} object to which the features have to be added.}
}
\description{
This is a function from Quintuitive that adds some features to a time series.
Actually the automatisms have been completely re-engineered and now all the
statistics refers to the actual day except for the \code{NEXTDAY} features
which indicates if the close price in the next trading day has been higher or
lower than the current price.
}
