% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_linear_trend}
\alias{add_linear_trend}
\title{Add Future Linear Trend}
\usage{
add_linear_trend(data, n = 10)
}
\arguments{
\item{data}{is an \code{xts} object to which the linear trend has to be
added.}

\item{n}{is the number of points for which the linear trend is suspected to
exist.}
}
\description{
This function adds to each point of a time series object the actual linear
trend in the immediate future.
}
