% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal.R
\docType{class}
\name{Deal}
\alias{Deal}
\title{Deal Class Object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Deal
}
\description{
The Deal Class creates an object that is able to handle long/short position
with stop-loss, take-profit and trailing stop features.
}
\keyword{datasets}
