% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-attempts.R
\name{test1}
\alias{test1}
\title{Test #1}
\usage{
test1()
}
\description{
This is a test function that launches some test code to test if this project
has any sense.
}
\details{
Jokes aside, it is the SVM example taken from Järjestelmällinen sijoittaja's
\href{https://jarjestelmallinensijoittaja.wordpress.com/tag/svm/}{article} on
Machine Learning in Time Series Prediction.
}
