% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-attempts.R
\name{trade_rsi}
\alias{trade_rsi}
\title{RSI Trading Signal}
\usage{
trade_rsi(series, lower = 30, upper = 70, epsilon = 5/100)
}
\arguments{
\item{series}{is a \code{xts} object representing the OHLC object as
retrieved using \code{getSymbols}.}

\item{lower}{is the RSI level requested to trigger the short trade.}

\item{upper}{is the RSI level requested to trigger the long trade.}

\item{epsilon}{is the tolerance to avoid closing a position too early.}
}
\description{
This is a strategy for long/short positioning according to RSI cross of
pre-fixed thresholds subject to parameterization.
}
